- Modified setup in touchDetect to actually fill the GUI from the internal settings during generation - Removed redundant Int conversion from keitley2612b.py

+++ plugins/Keithley2612B/Keithley2612B.py
@@ -271,7 +271,7 @@ class Keithley2612B:
             freq = 50
         else:
             freq = float(self.safequery("print(localnode.linefreq)"))
-            freq = int(round(freq))
+            freq = round(freq)
         return freq
 
     def getIV(self, channel) -> list[float]:
